requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
django-background-tasks==1.2.5
django-chartjs==2.2.1
django-compat==1.0.15
djangorestframework==3.12.2
et-xmlfile==1.0.1
flake8==3.8.4
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.5
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
